Revert "add transformation_id filter to jobs.list (#202)"
This reverts commit 7030a41702a495a20c1a05a58160f9d99872b699.
